Add STARTUP event dispatched by controllers.
This event is dispatched by a controller to its modules after all
of the controller's modules have been loaded, and should be used
for initialization that involves inter-module event dispatch.

Also fix the ordering of .start() calls on modules to respect priority,
and demote the notification of new events from warning to info.